requirements.txt
gspread==6.1.2
httpx==0.27.2
python-dotenv==1.0.1
google~=3.0.0